% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugrthesisdown.R
\docType{package}
\name{ugrthesisdown}
\alias{ugrthesisdown}
\alias{ugrthesisdown-package}
\title{ugrthesisdown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
Using the bookdown and thesisdown packages, and custom templates from
    Univeristy of Granada, this package aims to encourage reproducible research using
    simple Markdown syntax while embedding all of the R code to produce plots
    and analyses as well.
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document
}

\section{thesis_epub}{

Creates an R Markdown thesis template as an ebook
}

\author{
\strong{Maintainer}: Chester Ismay \email{chester.ismay@gmail.com}

Authors:
\itemize{
  \item Nick Solomon \email{nick@nicksolomon.me}
  \item Antonio J Pérez-Luque \email{ajpelu@gmail.com}
}

}
